requirements.txt
discord.py
python-dotenv
reportlab
